Fetch patient list in App.tsx
