+++ src/App.tsx
@@ -4,15 +4,28 @@ import { BrowserRouter as Router, Route, Link, Switch } from "react-router-dom";
 import { Button, Divider, Header, Container } from "semantic-ui-react";
 
 import { apiBaseUrl } from "./constants";
+import { useStateValue } from "./state";
+import { Patient } from "./types";
 
 import PatientListPage from "./PatientListPage";
 
 const App: React.FC = () => {
+  const [, dispatch] = useStateValue();
   React.useEffect(() => {
-    axios.get<void>(
-      `${apiBaseUrl}/ping`
-    );
-  }, []);
+    axios.get<void>(`${apiBaseUrl}/ping`);
+
+    const fetchPatientList = async () => {
+      try {
+        const { data: patientListFromApi } = await axios.get<Patient[]>(
+          `${apiBaseUrl}/patients`
+        );
+        dispatch({ type: "SET_PATIENT_LIST", payload: patientListFromApi });
+      } catch (e) {
+        console.error(e);
+      }
+    };
+    fetchPatientList();
+  }, [dispatch]);
 
   return (
     <div className="App">
